requirements.txt
python-telegram-bot
pydub
ffmpeg-python